[advanced] useNotification 훅 분리
